requirements.txt
Pillow==11.2.1
brokenaxes==0.6.2
h5py==3.14.0
matplotlib==3.10.3
networkx==3.5
numpy==2.3.1
opencv-python==4.11.0
pandas==2.3.0
scipy==1.16.0
tqdm==4.67.1
